Reorganized repo: moved notebooks, scripts, resources, docs, and data to new folders; removed empty files
