+++ data/demo.json
@@ -1,0 +1,5 @@
+{
+    "name": "Utkarsh Upadhyay",
+    "age": 20,
+    "gender": "male"
+}
