--- file/demo.json
@@ -1,5 +1,0 @@
-{
-    "name": "Utkarsh Upadhyay",
-    "age": 20,
-    "gender": "male"
-}
