+++ resources/Python Notes.pdf
