+++ resources/python cheat sheet.pdf
